agents/data-engineering/pyproject.toml
[project]
name = "data-engineering-agent"
version = "0.1.0"
description = "Data Engineering Agent - LLM Agent that can build Dataform pipelines."
authors = [
    { name = "Samet Karadag", email = "sametkaradag@google.com"},
    { name = "Saurabh Maurya", email = "saurabhmaurya@google.com"}
]
license = {text = "Apache License 2.0"}
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "env (>=0.1.0,<0.2.0)",
    "google-cloud-bigquery (>=3.11.4)",
    "google-cloud-dataform (>=0.4.0)",
    "google-api-core (>=2.15.0)",
    "google-auth (>=2.23.0)",
    "google-cloud-core (>=2.3.3)",
    "protobuf (>=4.24.4)",
    "typing-extensions (>=4.8.0)",
    "python-dotenv (>=1.0.0)",
    "google-adk (>=0.1.0)",
]

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"
black = "^25.1.0"
google-adk = { version = "^1.0.0", extras = ["eval"] }
pytest-asyncio = "^0.26.0"

[tool.poetry.group.deployment]
optional = true

[tool.poetry.group.deployment.dependencies]
absl-py = "^2.2.1"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
